Fix for 580 (#586)
* add lcz and altitude as attributes of site

* add site to api doc

* bugfix when altitude is nan

* set numeric to float64 for metadf

* fix tests part 1

* test fixing

* minor version bump

* black edits

* fix sphinx warnings

* code revieuw ready for merge

+++ tests/test_gee.py
@@ -146,10 +146,13 @@ class TestDemoDataset:
         # 5. Construct the equlity tests
         assert_equality(dataset, solutionobj)  # dataset comparison
 
-        assert_equality(alt_data["altitude"], solutionobj.metadf["altitude"])
+        assert_equality(
+            alt_data["altitude"].astype("float64"),
+            solutionobj.metadf["altitude"].astype("float64"),
+        )
         # calling printout
         _ = dataset.get_station("vlinder18").site.get_info(printout=False)
-        assert isinstance(dataset.get_station("vlinder18").site.altitude, int)
+        assert isinstance(dataset.get_station("vlinder18").site.altitude, float)
 
     def test_geemodeldata_getinfo_class(self):
         """Simple tests on GeeModelData instances"""
@@ -471,7 +474,7 @@ class TestDemoDataset:
 
 if __name__ == "__main__":
     test = TestDemoDataset()
-    # test.test_ERA5_extraction_on_metadata_only(overwrite_solution= False)
+    # test.test_ERA5_extraction_on_metadata_only(overwrite_solution=False)
     # test.test_import_demo_metadata(overwrite_solution=False)
     # test.test_LCZ_extraction(overwrite_solution=False)
     # test.test_altitude_extraction(overwrite_solution=False)
